### 2018-11-29 #1110 @whedon
[PRE REVIEW]: CEGO: C++11 Evolutionary Global Optimization
**Submitting author:** @ianhbell (<a href="http://orcid.org/0000-0003-1091-9080">Ian Bell</a>)
**Repository:** <a href="https://github.com/usnistgov/CEGO" target ="_blank">https://github.com/usnistgov/CEGO</a>
**Version:** v1.0.0
