### 2018-12-29 #1147 @whedon
[REVIEW]: CEGO: C++11 Evolutionary Globbal Optimization
**Submitting author:** @ianhbell (<a href="http://orcid.org/0000-0003-1091-9080">Ian Bell</a>)
**Repository:** <a href="https://github.com/usnistgov/CEGO" target ="_blank">https://github.com/usnistgov/CEGO</a>
**Version:** v1.0.0
