### 2018-12-29 #1147 @jedbrown
@sarats @sjvrijn @mmenickelly :wave: Welcome and thanks for agreeing to review! The comments from @whedon above outline the review process, which takes place in this thread (possibly with issues filed in the [CEGO repository](https://github.com/usnistgov/CEGO)). I'll be watching this thread if you have any questions.
