### 2019-01-05 #1147 @mmenickelly
Hi, first time reviewing for JOSS. To clarify this process: any comments I have that prevent me from checking off something in the checklist should be raised as an issue in the CEGO repository and not here, right?

And if I have multiple comments, is there a preference for separating them into multiple issues, or combining all comments into one large issue?
