### 2019-01-09 #1147 @sjvrijn
A few small comments so far @ianhbell:
 - The version in setup.py is still listed as 0.0.0 rather than 1.0.0
 - Installation worked fine through both methods, but I initially tried to install the listed dependencies separately before later realizing that they are automatically installed already. Might be worth mentioning in the Readme?
 - the doi for the 2006 paper is missing in the .bib file: 10.1145/1143997.1144142
