### 2019-02-27 #1147 @ianhbell
To everyone, sorry for the delay in dealing with this review...

@sjvrijn I will, as I have done for other projects, bump the revision to 1.0.0 once the review is complete, and at that time I will mint a DOI, and push to PyPI.  Thanks for the pointer about the DOI (fixed).  Didn't realize that the conference paper had a DOI.
