### 2019-02-27 23e70b0
Add doi for ALPS paper
