+++ JOSS/paper.bib
@@ -5,6 +5,7 @@
   title     = {{ALPS: The Age-Layered Population Structure for Reducing the Problem of Premature Convergence}},
   booktitle = {GECCO 2006},
   year      = {2006},
+  doi       = {10.1145/1143997.1144142}
 }
 
 @Conference{Hornby-GECCO-2009-ALPSsteady,
