### 2019-02-27 #1147 @ianhbell
Sorry for the delay everyone...

@sjvrijn Thanks for the pointer about the DOI; I fixed that, didn't know that the conference paper had a DOI.  I plan to bump the version to 1.0.0 once the review is complete, mint a DOI for the release, and push to PYPI.

@mmenickelly

1. Good point regarding docs.  I have set up to use doxygen to generate docs, and have committed a PDF of the documentation.  It's not a big file.
2. It's strictly unconstrained optimization.  Or rather, strictly softly constrained optimization with the ability to enforce integer constraints on values.  So no derivatives (which are a major obtain to generate in the general case).  I'll update the README.
3. You're right.  I thought I had set up Travis tests for CEGO, but I guess that was another project.  I'll do so and get them to fire on every commit.
4. That's odd.  For me, it yielded the global optimum on every run.  So I came in and ran it a few times, and indeed, it didn't always get there.  So I tweaked the control parameters a bit and added a caveat notice to each notebook along the lines of your comment

A1. I think so, I did something similar for ChebTools.  Do you think it is too much or too little?
A2. Fixed, see above
