+++ README.md
@@ -1,12 +1,17 @@
 # CEGO
 
-CEGO (pronounced sea-go) is C++11 Evolutionary Global Optimization.  It allows for:
+CEGO (pronounced *sea-go*) is C++11 Evolutionary Global Optimization.  This library is used to:
+
+* Solve unconstrained global optimization problems (Soft bound constraints can be added to the objective function as needed)
+* The input variables can be constrained to be integers (see [PressureVesselOptimization](https://nbviewer.jupyter.org/github/usnistgov/CEGO/tree/master/notebooks/PressureVesselOptimization.ipynb)  )
+
+It allows for:
 
 * A flexible C++11 architecture for doing parallel global optimization with multithreading
     * Also allows for new evolutionary optimization techniques to be specified with a minimum of code
     * Uses the age-layered approach
-* A C++ datatype (``CEGO::numberish``) that can be either an integer or a floating double precision value
-* Python wrappers of the core of the library
+* A fundamental C++ datatype (``CEGO::numberish``) that can be either an integer or a floating double precision value
+* Python wrappers of the core of the library (single-threaded)
 
 ## Examples:
 
@@ -19,6 +24,10 @@ Statically rendered examples are provided as Jupyter notebooks served on nbviewe
 * [PressureVesselOptimization](https://nbviewer.jupyter.org/github/usnistgov/CEGO/tree/master/notebooks/PressureVesselOptimization.ipynb) (MINLP optimization of pressure vessel mass)
 * [SpringOptimization](https://nbviewer.jupyter.org/github/usnistgov/CEGO/tree/master/notebooks/SpringOptimization.ipynb) (MINLP optimization of spring mass)
 * [Inverse Gaussian bumps](https://nbviewer.jupyter.org/github/usnistgov/CEGO/tree/master/notebooks/InverseBumps.ipynb) (inverse modeling with Gaussian bumps)
+
+## Documentation
+
+The C++ API documentation (generated by [doxygen](http://www.doxygen.nl/) ) is [available here](CEGO-1.0-doxygen.pdf)
 
 ## License
 
