+++ .travis.yml
@@ -1,0 +1,41 @@
+language: cpp
+compiler:
+  - clang
+  - gcc
+
+addons:
+  apt:
+    sources:
+    - george-edison55-precise-backports
+    - ubuntu-toolchain-r-test
+    - llvm-toolchain-precise-3.6
+    packages:
+    - cmake
+    - cmake-data
+    - gcc-5
+    - g++-5
+    - clang-3.6
+
+install:
+- if [ "$CXX" = "g++" ]; then export CXX="g++-5" CC="gcc-5"; fi
+- if [ "$CXX" = "clang++" ]; then export CXX="clang++-3.6" CC="clang-3.6"; fi
+
+# command to run tests
+script: 
+  - git submodule update --init
+  - mkdir build
+  - cd build
+  - cmake .. -DCEGO_ALL=ON
+  - cmake --build . --config Release
+  - ./Antoine
+  - ./fit_ratpoly_ancillary
+  - ./StornPriceProfiling
+  - ./main
+  - ./shaped_inverse_gaussian
+
+notifications:
+  email:
+    recipients:
+      - ian.bell@nist.gov
+    on_success: never # default: change
+    on_failure: always # default: always
