### 2019-02-27 5619d47
Don't build the Python interface on Travis
