+++ .travis.yml
@@ -25,7 +25,7 @@ script:
   - git submodule update --init
   - mkdir build
   - cd build
-  - cmake .. -DCEGO_ALL=ON
+  - cmake .. -DCEGO_ALL=ON -DCEGO_NO_PYTHON=ON
   - cmake --build . --config Release
   - ./Antoine
   - ./fit_ratpoly_ancillary
