### 2019-02-27 4d27a72
Add Travis badge
