+++ README.md
@@ -13,6 +13,10 @@ It allows for:
 * A fundamental C++ datatype (``CEGO::numberish``) that can be either an integer or a floating double precision value
 * Python wrappers of the core of the library (single-threaded)
 
+The C++ API documentation (generated by [doxygen](http://www.doxygen.nl/) ) is [available here](CEGO-1.0-doxygen.pdf)
+
+Automated Tests on TravisCI: [![Build Status](https://travis-ci.org/usnistgov/CEGO.svg?branch=master)](https://travis-ci.org/usnistgov/CEGO)
+
 ## Examples:
 
 Try it in your browser: [![Binder](https://mybinder.org/badge.svg)](https://mybinder.org/v2/gh/usnistgov/CEGO/master)
@@ -24,10 +28,6 @@ Statically rendered examples are provided as Jupyter notebooks served on nbviewe
 * [PressureVesselOptimization](https://nbviewer.jupyter.org/github/usnistgov/CEGO/tree/master/notebooks/PressureVesselOptimization.ipynb) (MINLP optimization of pressure vessel mass)
 * [SpringOptimization](https://nbviewer.jupyter.org/github/usnistgov/CEGO/tree/master/notebooks/SpringOptimization.ipynb) (MINLP optimization of spring mass)
 * [Inverse Gaussian bumps](https://nbviewer.jupyter.org/github/usnistgov/CEGO/tree/master/notebooks/InverseBumps.ipynb) (inverse modeling with Gaussian bumps)
-
-## Documentation
-
-The C++ API documentation (generated by [doxygen](http://www.doxygen.nl/) ) is [available here](CEGO-1.0-doxygen.pdf)
 
 ## License
 
