### 2019-02-27 #1147 @ianhbell
I have set up the Travis tests (https://travis-ci.org/usnistgov/CEGO), and pushed all changes to the repo.  Awaiting round number 2!
