### 2019-02-27 cdebc80
Try for a release build, maybe that would run faster on Travis
