### 2019-03-18 f3a7544
Also do the python build for sanity-checking of python compilation
