+++ .travis.yml
@@ -25,6 +25,7 @@ script:
   - git submodule update --init
   - mkdir build
   - cd build
+  - python setup.py build
   - cmake .. -DCEGO_ALL=ON -DCEGO_NO_PYTHON=ON -DCMAKE_BUILD_TYPE=Release
   - cmake --build . --config Release
   - ./StornPriceprofiling
