### 2019-03-18 a5c44f8
Make cmake happy on linux?
