+++ CMakeLists.txt
@@ -15,10 +15,10 @@ macro(add_nix_libraries target)
     FIND_PACKAGE ( Threads REQUIRED )
     find_package(Threads REQUIRED)
     if(THREADS_HAVE_PTHREAD_ARG)
-      target_compile_options(${target} PUBLIC "-pthread")
+      target_compile_options(${target} PRIVATE "-pthread")
     endif()
     if(CMAKE_THREAD_LIBS_INIT)
-      target_link_libraries(${target} "${CMAKE_THREAD_LIBS_INIT}")
+      target_link_libraries(${target} PRIVATE "${CMAKE_THREAD_LIBS_INIT}")
     endif()
 endmacro()
 
