+++ .travis.yml
@@ -23,9 +23,9 @@ install:
 # command to run tests
 script: 
   - git submodule update --init
+  - python setup.py build
   - mkdir build
   - cd build
-  - python setup.py build
   - cmake .. -DCEGO_ALL=ON -DCEGO_NO_PYTHON=ON -DCMAKE_BUILD_TYPE=Release
   - cmake --build . --config Release
   - ./StornPriceprofiling
