### 2019-03-18 dae9ac1
Revert "Try a clean Travis config.  Maybe that gives newer CMake?" .  Use standard Cmake version (3.x)
