+++ .travis.yml
@@ -5,12 +5,18 @@ compiler:
 
 addons:
   apt:
+    sources:
+    - george-edison55-precise-backports
+    - ubuntu-toolchain-r-test
+    - llvm-toolchain-precise-3.6
     packages:
-    - cmake
-    - cmake-data
-    - gcc
-    - g++
-    - clang
+    - gcc-5
+    - g++-5
+    - clang-3.6
+
+install:
+- if [ "$CXX" = "g++" ]; then export CXX="g++-5" CC="gcc-5"; fi
+- if [ "$CXX" = "clang++" ]; then export CXX="clang++-3.6" CC="clang-3.6"; fi
 
 # command to run tests
 script: 
