### 2019-03-18 #1147 @ianhbell
I have fixed the build issues with binder and now build the Python wrapper in TravisCI as well.  All's well at the moment, ready for another look-see.
