### 2019-04-01 #1147 @sjvrijn
A few more remarks:

- Could you expand the Community Guidelines a bit more? They do mention contributions/bugs, but regular help with using the package is not addressed

- Notebooks seem to work as intended, but are not consistent in their presentation. I don't mind it that much, but at least explicitly listing the intended global optimum and a reference for the used function would be nice

    - `SpringOptimization.ipynb`
        - no source for problem is listed
        - it is unclear to me what the four printed values below cell 2 mean
        - the result using `pyswarms` in cell 4 seems to be better than yours. Is that expected behavior?
    - `NonIntegerConstraing.ipynb`
        - add a link to the given source
    - `Griewank.ipynb`
        - list expected optimum to be 0

PS: the use of `for i in range(len(x))` in the Griewank definition is not wrong, but more 'pythonic' would be:
```python
    for i, xi in enumerate(x):
        sum1 += pow(xi.as_double(), 2)/4000.0
        prod1 *= cos(xi.as_double()/sqrt(i+1))
```
🙂
