+++ README.md
@@ -42,7 +42,9 @@ Statically rendered examples are provided as Jupyter notebooks served on nbviewe
 
 ## Contributing/Getting Help
 
-If you would like to contribute to ``CEGO`` or report a problem, please open a pull request or submit an issue.  Especially welcome would be additional tests.
+If you would like to contribute to ``CEGO`` or report a problem, please open a pull request or submit an issue.  Especially welcome would be additional tests.  
+
+To get started, you should check out the Jupyter notebooks; they demonstrate some of the capabilities of CEGO.
 
 ## Installation
 
