### 2019-04-05 #1147 @ianhbell
Thanks @sjvrijn -- those were all great recommendations.  I have made all them and am ready for another look-see.  Some comments:

0) I wasn't entirely clear what you had in mind here, so I added ```To get started, you should check out the Jupyter notebooks in the notebooks folder; they demonstrate some of the capabilities of CEGO.```
1) The call to pyswarms doesn't satisfy the constraints since pyswarms only does continuous optimization
2) Thanks for the pythonic suggestion.  It's a pity that they don't do a better job talking about the enumerate function in the docs of Python.  It's only much later on I learned about this very useful function.  Indexing ``x`` twice is not a huge penalty time-wise, but I agree your version is nicer to read and saves two characters per line
