### 2019-04-22 db4aab2
JOSS: minor editing and bib fixes; add Jupyter and Binder citations
https://github.com/jupyter/jupyter/issues/190
https://mybinder.readthedocs.io/en/latest/#citing-binder
