### 2019-04-22 #1147 @jedbrown
Can you please merge the above PR and comment on the 2009 paper that seems to have not been published in the stated venue? Then we can double-check the paper before archiving. Thanks.
