### 2019-04-22 d99957e
Remove the mysterious Hornby paper from 2009
