### 2019-04-22 #1147 @ianhbell
Hornby paper has been removed in https://github.com/usnistgov/CEGO/commit/d99957e2ffb65d0d632a94c15385e752ac70368a
