### 2019-04-23 #1147 @jedbrown
@ianhbell Looks good. Please tag a release ([annotated tags](https://git-scm.com/book/en/v2/Git-Basics-Tagging#_annotated_tags) preferred), archive on Zenodo or similar, and report the DOI here.
