### 2019-04-23 #1147 @danielskatz
The DOI resolves correctly, and the landing page for the paper is fine, but the PDF of the paper is 404 for me
