### 2019-04-23 #1147 @danielskatz
@arfon, any thoughts on this before I close the issue?  I'll also try again in 2 hours from another location and see if it works there
