### 2019-04-23 #1147 @arfon
> Now it’s working for me

The PDFs are cached in Cloudflare which means that if you click the link to the PDF before GitHub pages has had a chance to build (and deploy) then you can get a 404. Your browser (and Cloudflare) can then sometimes cache the 404 response (I think).
